+++ src/lib/index.ts
@@ -1,2 +1,8 @@
-export { Incident, format } from "./incident";
+import { format, Incident } from "./incident";
+
+export { format, Incident } from "./incident";
 export { StaticIncident } from "./types";
+
+const DEFAULT: any = {format, Incident};
+
+export { DEFAULT as default };